Fixed upload-file tmp file space leak
Cherry picked from https://github.com/ckan/ckan/issues/3752 (with modifications)

Moved common_middleware.py and pylons_app.py changes to middleware.py

Updated "if upload:" to "if upload is not None:" because otherwise it
evaluated to false and skipped the "upload.fp.close()". This change was
added to the ckan PR by smotornyuk.

https://trello.com/c/oP4frPB6/200-very-large-files-cannot-be-added-to-open-government-portal