Fix broken upload-file fix
Fix 1

For small files, self.upload_file in uploader.py is a StringO
(cStringIO.OutputType) not a file so "with:" fails. Replaced "with:"
with a self.upload_file.close(). In this 2.3.5 version, I accidentally
grabbed the old version of the fix from the ckan people before they
made this exact fix.

Put this upload_file.close() in a try-finally block so that any
exceptions (eg the file size exceeds the max), will close the file
and not cause a tmp space leak.

Fix 2

For resources which are a link, "upload" in middleware.py is type
string, not type cgi.FieldStorage. It is actually an empty string.
It can also be "None" for regular non-upload http requests. So now
before closing upload.fp, we make sure that "upload" implements "fp".
Note that "if upload:" is false for type FieldStorage.

https://trello.com/c/mCTGrKdu/517-fix-large-file-upload-errors